unify vlm dataset format
