+++ xtuner/v1/data_proto/templates/__init__.py
@@ -13,6 +13,9 @@ CHAT_TEMPLATE_MAP = {
         user="<|im_start|>user\n{user}<|im_end|>\n<|im_start|>assistant\n",
         assistant="{assistant}<|im_end|>",
         stop_words=["<|im_end|>"],
+        image_start_token="<img>",
+        image_end_token="</img>",
+        image_context_token="<IMG_CONTEXT>",
     ),
     "llama3": HybridChatTemplate(
         system=("<|start_header_id|>system<|end_header_id|>\n\n{system}<|eot_id|>"),
