+++ xtuner/v1/data_proto/templates/hybrid.py
@@ -25,7 +25,10 @@ class HybridChatTemplate(BaseModel):
 
     # Multimodal Chat
     # Predefined token and index for images
-    image_token: str = "<image>"
+    image_context_token: str | None = None
+    image_token: str = "<IMAGE_TOKEN>"
+    image_start_token: str = ""
+    image_end_token: str = ""
     image_token_index: int = -100
 
     # Agent Chat
