--- xtuner/v1/datasets/intern_s1_fn/process.py
@@ -1,349 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
-import io
-from typing import Dict, List, cast
-
-import numpy as np
-import torch
-import torchvision.transforms as T
-from PIL import Image
-from torchvision.transforms.functional import InterpolationMode
-
-import transformers
-from transformers.trainer_pt_utils import LabelSmoother
-
-
-IGNORE_TOKEN_ID = LabelSmoother.ignore_index
-IMG_CONTEXT_TOKEN = "<IMG_CONTEXT>"
-IMG_START_TOKEN = "<img>"
-IMG_END_TOKEN = "</img>"
-QUAD_START_TOKEN = "<quad>"
-QUAD_END_TOKEN = "</quad>"
-REF_START_TOKEN = "<ref>"
-REF_END_TOKEN = "</ref>"
-BOX_START_TOKEN = "<box>"
-BOX_END_TOKEN = "</box>"
-IMAGENET_MEAN = (0.485, 0.456, 0.406)
-IMAGENET_STD = (0.229, 0.224, 0.225)
-CLIP_MEAN = (0.4814546, 0.4578275, 0.40821073)
-CLIP_STD = (0.2686295, 0.2613025, 0.2757711)
-SIGLIP_MEAN = (0.5, 0.5, 0.5)
-SIGLIP_STD = (0.5, 0.5, 0.5)
-IGNORE_INDEX = -100
-
-
-def expand2square(pil_img, background_color):
-    width, height = pil_img.size
-    if width == height:
-        return pil_img
-    elif width > height:
-        result = Image.new(pil_img.mode, (width, width), background_color)
-        result.paste(pil_img, (0, (width - height) // 2))
-        return result
-    else:
-        result = Image.new(pil_img.mode, (height, height), background_color)
-        result.paste(pil_img, ((height - width) // 2, 0))
-        return result
-
-
-def simulate_jpeg_degradation(quality):
-    def jpeg_degrade(img):
-        with io.BytesIO() as output:
-            img.convert("RGB").save(output, format="JPEG", quality=quality)
-            output.seek(0)  # Move the reading cursor to the start of the stream
-            img_jpeg = Image.open(output).copy()  # Use .copy() to make sure the image is loaded in memory
-        return img_jpeg
-
-    return jpeg_degrade
-
-
-# Define the JPEG compression quality range, pre-create all JPEG compression functions
-qualities = list(range(75, 101))
-jpeg_degrade_functions = {quality: simulate_jpeg_degradation(quality) for quality in qualities}
-
-
-def build_transform(is_train, input_size, pad2square=False, normalize_type="imagenet"):
-    if normalize_type == "imagenet":
-        MEAN, STD = IMAGENET_MEAN, IMAGENET_STD
-    elif normalize_type == "clip":
-        MEAN, STD = CLIP_MEAN, CLIP_STD
-    elif normalize_type == "siglip":
-        MEAN, STD = SIGLIP_MEAN, SIGLIP_STD
-    else:
-        raise NotImplementedError
-    if is_train:  # use data augumentation
-        transform = T.Compose(
-            [
-                T.Lambda(lambda img: img.convert("RGB") if img.mode != "RGB" else img),
-                T.RandomChoice([T.Lambda(jpeg_degrade_functions[quality]) for quality in qualities]),
-                T.Resize((input_size, input_size), interpolation=InterpolationMode.BICUBIC),
-                T.ToTensor(),
-                T.Normalize(mean=MEAN, std=STD),
-            ]
-        )
-    else:
-        if pad2square is False:  # now we use this transform function by default
-            transform = T.Compose(
-                [
-                    T.Lambda(lambda img: img.convert("RGB") if img.mode != "RGB" else img),
-                    T.Resize(
-                        (input_size, input_size),
-                        interpolation=InterpolationMode.BICUBIC,
-                    ),
-                    T.ToTensor(),
-                    T.Normalize(mean=MEAN, std=STD),
-                ]
-            )
-        else:
-            transform = T.Compose(
-                [
-                    T.Lambda(lambda img: img.convert("RGB") if img.mode != "RGB" else img),
-                    T.Lambda(lambda img: expand2square(img, tuple(int(x * 255) for x in MEAN))),
-                    T.Resize(
-                        (input_size, input_size),
-                        interpolation=InterpolationMode.BICUBIC,
-                    ),
-                    T.ToTensor(),
-                    T.Normalize(mean=MEAN, std=STD),
-                ]
-            )
-
-    return transform
-
-
-def preprocess_intern_s1(
-    sources,
-    tokenizer: transformers.PreTrainedTokenizer,
-    num_image_token_list: list,
-    text_only: bool = False,
-    ds_name: str = None,  # type: ignore
-    num_image: int = 1,
-    prompt_only: bool = False,
-    system_prompt: str = None,  # type: ignore
-    max_length: int | None = None,
-) -> Dict[str, list[int]]:
-    if max_length is not None:
-        tokenizer.model_max_length = max_length
-    assert len(sources) == 1, "ERROR: process only the first conversations"
-    conversations = sources[0]
-
-    # process pretrain data
-    if conversations[0]["from"] == "pretrain":
-        assert len(conversations) == 1, "ERROR: pretrain data should only have one message"
-        input_text = conversations[0]["value"]
-        if not text_only:
-            image_cnt = input_text.count("<image>")
-            assert image_cnt == num_image, (
-                f"ERROR: {ds_name}, image_cnt: {image_cnt} != num_image: {num_image}, {input_text}"
-            )
-            for i in range(num_image):
-                image_tokens = f"{IMG_START_TOKEN}{IMG_CONTEXT_TOKEN * num_image_token_list[i]}{IMG_END_TOKEN}"
-                input_text = input_text.replace("<image>", image_tokens, 1)
-        input_ids = tokenizer.encode(input_text, add_special_tokens=False)
-        labels = input_ids.copy()
-        # ignore loss for image tokens
-        img_start_token_id = tokenizer.convert_tokens_to_ids(IMG_START_TOKEN)
-        img_context_token_id = tokenizer.convert_tokens_to_ids(IMG_CONTEXT_TOKEN)
-        img_end_token_id = tokenizer.convert_tokens_to_ids(IMG_END_TOKEN)
-        labels[labels == img_start_token_id] = -100
-        labels[labels == img_context_token_id] = -100
-        labels[labels == img_end_token_id] = -100
-
-        if len(input_ids) > tokenizer.model_max_length:
-            print(
-                f"WARNING: input_ids length {len(input_ids)} exceeds "
-                f"model_max_length {tokenizer.model_max_length}. truncated!"
-            )
-            input_ids = input_ids[: tokenizer.model_max_length]
-            labels = labels[: tokenizer.model_max_length]
-
-        input_ids = torch.tensor(input_ids, dtype=torch.long)
-        labels = torch.tensor(labels, dtype=torch.long)
-
-        if not text_only:
-            assert (input_ids == img_end_token_id).sum() == num_image, (
-                f"ERROR: image tokens are truncated, this dataset is {ds_name}"
-            )
-
-        assert input_ids.size() == labels.size()
-
-        input_ids = input_ids.unsqueeze(0)
-        labels = labels.unsqueeze(0)
-
-        return dict(input_ids=input_ids.tolist(), labels=labels.tolist())
-
-    if conversations[0]["from"] == "system":
-        system_prompt = conversations[0]["value"]
-        conversations = conversations[1:]  # remove system prompt
-    else:
-        system_prompt = system_prompt
-
-    if not text_only:
-        new_conversations = []
-        current_image_idx = 0
-        for conversation in conversations:
-            if conversation["from"] == "human":
-                image_cnt = conversation["value"].count("<image>")
-                for i in range(image_cnt):
-                    if current_image_idx == num_image:
-                        break
-                    image_tokens = f"{IMG_START_TOKEN}{IMG_CONTEXT_TOKEN * num_image_token_list[current_image_idx]}{IMG_END_TOKEN}"
-                    conversation["value"] = conversation["value"].replace("<image>", image_tokens, 1)
-                    current_image_idx += 1
-            new_conversations.append(conversation)
-        conversations = new_conversations
-        # if current_image_idx < num_image, it means <image> placeholder is less than num_image
-        assert current_image_idx == num_image, (
-            f"ERROR: current_image_idx: {current_image_idx} != num_image: {num_image}"
-        )
-
-    batches, roles = [], []
-    if system_prompt is not None:
-        batches.append(f"<|im_start|>system\n{system_prompt}<|im_end|>\n")
-        roles.append("system")
-    if not prompt_only:
-        for conversation in conversations:
-            if conversation["from"] == "human":
-                batches.append(f"<|im_start|>user\n{conversation['value']}<|im_end|>\n")
-                roles.append("human")
-            elif conversation["from"] == "gpt":
-                batches.append(f"<|im_start|>assistant\n{conversation['value']}<|im_end|>\n")
-                roles.append("gpt")
-            elif conversation["from"] == "function":
-                batches.append(f"<|im_start|>function\n{conversation['value']}<|im_end|>\n")
-                roles.append("function")
-            else:
-                raise ValueError(f"ERROR role: {conversation['from']}: dataset name: {ds_name}")
-    else:
-        assert conversations[0]["from"] == "human", (
-            "ERROR: prompt_only is True, the first message should be from human"
-        )
-        batches.append(f"<|im_start|>user\n{conversations[0]['value']}<|im_end|>\n")
-        roles.append("human")
-
-    input_ids = tokenizer(
-        batches,
-        return_tensors="np",
-        padding=False,
-        truncation=False,
-    ).input_ids
-
-    final_input_ids, final_targets = [], []
-    ignore_ids = tokenizer("<|im_start|>assistant\n", return_tensors="np").input_ids[0]
-    ignore_len = ignore_ids.shape[0]
-    for role, input_id in zip(roles, input_ids):
-        final_input_ids.append(input_id)
-        if role == "system" or role == "human" or role == "function":
-            final_targets.append(np.full(input_id.shape, IGNORE_TOKEN_ID))  # ignore
-        elif role == "gpt":
-            target = input_id.copy()
-            target[:ignore_len] = IGNORE_TOKEN_ID  # ignore loss for `<|im_start|>assistant\n`
-            target[-1:] = IGNORE_TOKEN_ID  # ignore loss for `\n`
-            final_targets.append(target)
-        else:
-            raise ValueError(f"ERROR role: {role}: dataset name: {ds_name}")
-
-    input_ids = np.concatenate(final_input_ids)
-    targets = np.concatenate(final_targets)
-
-    if len(input_ids) > tokenizer.model_max_length:
-        print(
-            f"WARNING: input_ids length {len(input_ids)} exceeds "
-            f"model_max_length {tokenizer.model_max_length}. truncated!"
-        )
-        input_ids = input_ids[: tokenizer.model_max_length]
-        targets = targets[: tokenizer.model_max_length]
-
-    if not text_only:
-        image_end_token_id = tokenizer.convert_tokens_to_ids(IMG_END_TOKEN)
-        assert (input_ids == image_end_token_id).sum() == num_image, (
-            f"ERROR: image tokens are truncated, this dataset is {ds_name}"
-        )
-
-    return dict(
-        input_ids=cast(List[int], input_ids.tolist()),
-        labels=cast(List[int], targets.tolist()),
-    )
-
-
-def find_closest_aspect_ratio(aspect_ratio, target_ratios, width, height, image_size):
-    best_ratio_diff = float("inf")
-    best_ratio = (1, 1)
-    area = width * height
-    for ratio in target_ratios:
-        target_aspect_ratio = ratio[0] / ratio[1]
-        ratio_diff = abs(aspect_ratio - target_aspect_ratio)
-        if ratio_diff < best_ratio_diff:
-            best_ratio_diff = ratio_diff
-            best_ratio = ratio
-        elif ratio_diff == best_ratio_diff:
-            if area > 0.5 * image_size * image_size * ratio[0] * ratio[1]:
-                best_ratio = ratio
-    # print(f'width: {width}, height: {height}, best_ratio: {best_ratio}')
-    return best_ratio
-
-
-def dynamic_preprocess(image, min_num=1, max_num=6, image_size=448, use_thumbnail=False):
-    orig_width, orig_height = image.size
-    aspect_ratio = orig_width / orig_height
-
-    # calculate the existing image aspect ratio
-    target_ratios = {
-        (i, j)
-        for n in range(min_num, max_num + 1)
-        for i in range(1, n + 1)
-        for j in range(1, n + 1)
-        if max_num >= i * j >= min_num
-    }
-    target_ratios = sorted(target_ratios, key=lambda x: x[0] * x[1])
-
-    # find the closest aspect ratio to the target
-    target_aspect_ratio = find_closest_aspect_ratio(aspect_ratio, target_ratios, orig_width, orig_height, image_size)
-
-    # calculate the target width and height
-    target_width = image_size * target_aspect_ratio[0]
-    target_height = image_size * target_aspect_ratio[1]
-    blocks = target_aspect_ratio[0] * target_aspect_ratio[1]
-
-    # resize the image
-    resized_img = image.resize((target_width, target_height))
-    processed_images = []
-    for i in range(blocks):
-        box = (
-            (i % (target_width // image_size)) * image_size,
-            (i // (target_width // image_size)) * image_size,
-            ((i % (target_width // image_size)) + 1) * image_size,
-            ((i // (target_width // image_size)) + 1) * image_size,
-        )
-        # split the image
-        split_img = resized_img.crop(box)
-        processed_images.append(split_img)
-    assert len(processed_images) == blocks
-    if use_thumbnail and len(processed_images) != 1:
-        thumbnail_img = image.resize((image_size, image_size))
-        processed_images.append(thumbnail_img)
-    return processed_images
-
-
-def dynamic_num_patch(size, min_num=1, max_num=6, image_size=448, use_thumbnail=False):
-    orig_width, orig_height = size
-    aspect_ratio = orig_width / orig_height
-
-    # calculate the existing image aspect ratio
-    target_ratios = {
-        (i, j)
-        for n in range(min_num, max_num + 1)
-        for i in range(1, n + 1)
-        for j in range(1, n + 1)
-        if i * j <= max_num and i * j >= min_num
-    }
-    target_ratios = sorted(target_ratios, key=lambda x: x[0] * x[1])
-
-    # find the closest aspect ratio to the target
-    target_aspect_ratio = find_closest_aspect_ratio(aspect_ratio, target_ratios, orig_width, orig_height, image_size)
-
-    # calculate the target width and height
-    blocks = target_aspect_ratio[0] * target_aspect_ratio[1]
-
-    if use_thumbnail and blocks > 1:
-        blocks += 1
-    return blocks
