+++ xtuner/v1/datasets/mllm_tokenize_fn/__init__.py
@@ -1,0 +1,7 @@
+from .intern_s1_vl_tokenize_fn import InternS1VLTokenizeFnConfig, InternS1VLTokenizeFunction
+
+
+__all__ = [
+    "InternS1VLTokenizeFunction",
+    "InternS1VLTokenizeFnConfig",
+]
